Fix resolving an image for captioning.
Do *not* use restrictedTraverse here. Any folder along the resolving path could lack the View permission for the current visitor (could be Anonymous), resulting in a authorization challenge.

Note that we do not validate the image itself for the View permission either; we rather show a broken image icon in the browser than raise Unauthorized here or not render the image.