Update plone/outputfilters/filters/resolveuid_and_caption.py
I am upgrading an old Plone 2.1 site to 4.2. Using Kupu or (F)CKeditor or whatever with the UID link option turned on left me with a lot of pages having links to "resolveUid/..." instead of the current "resolveuid/..." (note the upper-case "U"). This patch makes the old links work again.