Revert "Add some deprecation warnings in resolveuid."
This reverts commit 6ee71503c52c422a795fba74625b46f3f6a553df.
It causes test failures in plone.restapi.

Also, we should test what happens in our own code when you are anonymous and we call `uuidToObject` on a uuid of a private object.
Actually, that seems to work, at least for a simple link, both with our version of this function and with the version from `plone.app.uuid`.
But I don't know why it works.