Removed _shorttag_replace from picture_variants filter.
This was copied from the resolve_uid_and_caption filter, but was missing the singleton_tags definition.

I don't think it is needed in resolve_uid_and_caption either.
It is meant to replace singleton tags without any content which should not be singleton tags.
For example change `<p />` into `<p></p>`.
But beautifulsoup is already doing that for us:

```
>>> print(BeautifulSoup("<p />", "html.parser"))
<p></p>
```

Cleanup (plus possibly adding a test) can be done later, to not overfill this PR.